Save work on reducing warnings

+++ src/x86/memory.cpp
@@ -206,7 +206,7 @@ uint32_t Free86::ld16_direct(uint32_t address) {
     return memory[address] | (memory[address + 1] << 8);
 }
 uint32_t Free86::ld_direct(uint32_t address) {
-    return memory[address] | (memory[address + 1] << 8) | (memory[address + 2] << 16) | (memory[address + 3] << 24);
+    return memory[address] | (static_cast<uint32_t>(memory[address + 1]) << 8) | (static_cast<uint32_t>(memory[address + 2]) << 16) | (static_cast<uint32_t>(memory[address + 3]) << 24);
 }
 uint64_t Free86::ld64_direct(uint32_t address) {
     return (ld_direct(address) & 0xffffffff) | (static_cast<uint64_t>(ld_direct(address + 4)) << 32);
